Use Akavache for caching the list of cities
Use Akavache for caching the list of cities. This commit pulls in a
bunch of external dependencies that are required by Akavache. Also, if
you build a Windows app, you need to add the SQLite for Windows
component to your WinPhone project (look for it in Visual Studio
Extensions) since Akavache relies on SQLite for on device storage.