Use ModernHttpClient for native network stack
Use ModernHttpClient for native networking. Please note that this commit
deliberately breaks the iOS app, to illustrate that the network requests
are actually going through the native network stack. Uncomment the
NSAppTransportSecurity setting in the info.plist file to work around the
App Transport Security requirement that was added since iOS 9 by Apple.